confidential: move commitment types into their own modules
The code for value, nonce, asset, have diverged in various ways. By splitting
them into multiple files I hope to make diffing them easier.

Code move only; the next commits will make chonges to bring the three files
closer together.

Also I am quietly relicensing this stuff from CC0 to MIT+Apache. I am not
doing this to be sneaky; I was just creating new files so it was an opportunity
to use a new license header.
See https://github.com/rust-bitcoin/rust-bitcoin/issues/5849
